PID Auto Align (#58)
Use a PID Controller to aim the robot at the target. This is much faster than the previous method of aiming which was only proportional.

The motors will not turn if the value wrtten to them is in the range +-0.3. I added a base turn speed of 0.3 to the PID output so that any change will have an immediate effect on the motors.